+++ @remirror/extension-code-block/src/code-block-utils.ts
@@ -223,7 +223,7 @@ const PRELOADED_LANGUAGES = [markup, clike, css, js];
  * The list of strings that are recognised language names based on the the configured
  * supported languages.
  */
-export const getLanguageNamesAndAliases = (supportedLanguages: RefractorSyntax[]) => {
+export const getLanguageNamesAndAliases = (supportedLanguages: RefractorSyntax[]): string[] => {
   return uniqueArray(
     flattenArray(
       [...PRELOADED_LANGUAGES, ...supportedLanguages].map(({ name, aliases }) => [
@@ -232,13 +232,6 @@ export const getLanguageNamesAndAliases = (supportedLanguages: RefractorSyntax[]
       ]),
     ),
   );
-};
-
-/**
- * Returns true if the language is supported.
- */
-export const isSupportedLanguage = (language: string, supportedLanguages: RefractorSyntax[]) => {
-  return getLanguageNamesAndAliases(supportedLanguages).includes(language);
 };
 
 interface GetLanguageParameter {
@@ -261,8 +254,21 @@ interface GetLanguageParameter {
 /**
  * Get the language from user input.
  */
-export const getLanguage = ({ language, supportedLanguages, fallback }: GetLanguageParameter) =>
-  !isSupportedLanguage(language, supportedLanguages) ? fallback : language;
+export const getLanguage = (parameter: GetLanguageParameter): string => {
+  const { language, supportedLanguages, fallback } = parameter;
+
+  if (!language) {
+    return fallback;
+  }
+
+  for (const name of getLanguageNamesAndAliases(supportedLanguages)) {
+    if (name.toLowerCase() === language.toLowerCase()) {
+      return name;
+    }
+  }
+
+  return fallback;
+};
 
 interface FormatCodeBlockFactoryParameter
   extends NodeTypeParameter,
@@ -336,18 +342,4 @@ export const formatCodeBlockFactory = ({
   }
 
   return true;
-};
-
-/**
- * Retrieve the supported language names based on configuration.
- */
-export const getSupportedLanguagesMap = (supportedLanguages: RefractorSyntax[]) => {
-  const object_: Record<string, string> = object();
-  for (const { name, aliases } of [...PRELOADED_LANGUAGES, ...supportedLanguages]) {
-    object_[name] = name;
-    aliases.forEach((alias) => {
-      object_[alias] = name;
-    });
-  }
-  return object_;
 };