+++ @remirror/editor-wysiwyg/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @remirror/editor-wysiwyg
 
+## 0.13.0
+
+### Minor Changes
+
+- 70ab69c8: Change import of w3c-keyname@2 from default to named due to breaking change in package.
+
+## 0.12.0
+
+### Patch Changes
+
+- Updated dependencies [a43b0fd5]
+- Updated dependencies [44f85e79]
+  - @remirror/extension-code-block@0.12.0
+
 ## 0.11.0
 
 ### Patch Changes