+++ docs/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @remirror/docs
 
+## 0.0.5
+
+### Patch Changes
+
+- @remirror/showcase@0.13.0
+
+## 0.0.4
+
+### Patch Changes
+
+- @remirror/showcase@0.12.0
+
 ## 0.0.3
 
 ### Patch Changes