+++ examples/with-razzle/CHANGELOG.md
@@ -1,5 +1,17 @@
 # with-razzle
 
+## 0.0.3
+
+### Patch Changes
+
+- @remirror/showcase@0.13.0
+
+## 0.0.2
+
+### Patch Changes
+
+- @remirror/showcase@0.12.0
+
 ## 0.0.1
 
 ### Patch Changes